+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.2.0
+ - **FEAT**: Add disabled buttons to controller
+ - **FEAT**: Add controller fields initialization in controller constructor
+ - **INFO**: Add deprecatd annotation for selectedButton, selectedButtons and disabledButtons
+
 ## 4.1.1
  - **FIX**: GroupbuttonController initialization
 ## 4.1.0